xview2-model/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.25.0
numpy>=1.21.0
